pyproject.toml
[project]
name = "videoshelf"
version = "0.1.0"
description = "Video processing and cutting tool"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "celery[redis]",
    "redis",
    "ffmpeg-python",
    "boto3",
    "pydantic>=2.0",
    "pydantic-settings",
    "python-multipart",
    "websockets",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "watchdog",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
